move methods away from Live controller
