+++ plugins/Live/Live.php
@@ -8,6 +8,14 @@
  */
 namespace Piwik\Plugins\Live;
 
+use Piwik\Cache;
+use Piwik\CacheId;
+use Piwik\API\Request;
+use Piwik\Common;
+use Piwik\Plugin;
+use Piwik\Piwik;
+use Piwik\Plugins\Live\ProfileSummary\ProfileSummaryAbstract;
+
 /**
  *
  */
@@ -113,5 +121,92 @@ class Live extends \Piwik\Plugin
         foreach ($visitorDetailsInstances as $instance) {
             $renderedDetails .= $instance->renderIcons($visitorDetails);
         }
+    }
+
+    public static function getSegmentWithVisitorId()
+    {
+        static $cached = null;
+        if ($cached === null) {
+            $segment = Request::getRawSegmentFromRequest();
+            if (!empty($segment)) {
+                $segment = urldecode($segment) . ';';
+            }
+
+            $idVisitor = Common::getRequestVar('visitorId', false);
+            if ($idVisitor === false) {
+                $idVisitor = Request::processRequest('Live.getMostRecentVisitorId');
+            }
+
+            $cached = urlencode($segment . 'visitorId==' . $idVisitor);
+        }
+        return $cached;
+    }
+
+
+    /**
+     * Returns all available profile summaries
+     *
+     * @return ProfileSummaryAbstract[]
+     * @throws \Exception
+     */
+    public static function getAllProfileSummaryInstances()
+    {
+        $cacheId = CacheId::pluginAware('ProfileSummaries');
+        $cache   = Cache::getTransientCache();
+
+        if (!$cache->contains($cacheId)) {
+            $instances = [];
+
+            /**
+             * Triggered to add new live profile summaries.
+             *
+             * **Example**
+             *
+             *     public function addProfileSummary(&$profileSummaries)
+             *     {
+             *         $profileSummaries[] = new MyCustomProfileSummary();
+             *     }
+             *
+             * @param ProfileSummaryAbstract[] $profileSummaries An array of profile summaries
+             */
+            Piwik::postEvent('Live.addProfileSummaries', array(&$instances));
+
+            foreach (self::getAllProfileSummaryClasses() as $className) {
+                $instances[] = new $className();
+            }
+
+            /**
+             * Triggered to filter / restrict profile summaries.
+             *
+             * **Example**
+             *
+             *     public function filterProfileSummary(&$profileSummaries)
+             *     {
+             *         foreach ($profileSummaries as $index => $profileSummary) {
+             *              if ($profileSummary->getId() === 'myid') {}
+             *                  unset($profileSummaries[$index]); // remove all summaries having this ID
+             *              }
+             *         }
+             *     }
+             *
+             * @param ProfileSummaryAbstract[] $profileSummaries An array of profile summaries
+             */
+            Piwik::postEvent('Live.filterProfileSummaries', array(&$instances));
+
+            $cache->save($cacheId, $instances);
+        }
+
+        return $cache->fetch($cacheId);
+    }
+
+    /**
+     * Returns class names of all VisitorDetails classes.
+     *
+     * @return string[]
+     * @api
+     */
+    protected static function getAllProfileSummaryClasses()
+    {
+        return Plugin\Manager::getInstance()->findMultipleComponents('ProfileSummary', 'Piwik\Plugins\Live\ProfileSummary\ProfileSummaryAbstract');
     }
 }
